Re-layout list editor
The aim of this PR is to improve the UX of the list editor.

- The editor now has a left hand side with all items tagged to the
item. On the right all lists are displayed. This hopefully improves the
understanding of the relationship between the two parts.
- Items from the left hand side are only "draggable". Previously it was
possible to sort and drag items in this list, causing confusion. It
also allows users to add a link to multiple lists, something which
could previously only be achieved by turning off javascript.
- When an item is included in one or more lists it is bold
- When the user is about to create the first list for a tag we show a
special page explaining they're about to change the "sort mode" of the
tag. This can be further improved by setting the ordering explicitly -
but that's for a later PR.
- The layout of the tables is updated to fit more items on the screen
and make dragging easier.
- Previously, when an item was dragged we calculated the exact rows of
that had to be updated. For example, when the user dragged the fifth
item in a list of 10 to the fourth place, we would update items 4
through 10. Finding out which items to update turned out to be a
fruitful source of bugs and confusion. This behaviour has been changed
to always update all items in a list on each update. In the previous
example, instead of 4-10, we update 1-10. In a later PR, we should use
bulk-updating to the server (which is now not very hard to implement).
- Related, keeping data in sync is hard on JS.
GOVUK.curatedLists.showEmptyLabels and
GOVUK.curatedLists.redrawUncategorizedList are purposely "naive". They
look at the state of the world and re-render it on each update, instead
of trying to only update things that changed, which would be more
performant.

+++ app/assets/javascripts/curated_lists.js
@@ -5,91 +5,52 @@
 
   GOVUK.curatedLists = {
     init: function() {
-      var $lists = $('.curated-list');
+      GOVUK.curatedLists.redrawUncategorizedList();
+      GOVUK.curatedLists.activateRemoveButtons();
+      GOVUK.curatedLists.showEmptyLabels();
 
-      $lists.each(function(_, list) {
-        var $list = $(list);
-
-        if ($list.children().length > 1) {
-          $list.children('.empty-list').hide();
-        }
-
-        $list.sortable({
-          connectWith: '.curated-list',
-          dropOnEmpty: true,
-          stop: function(event, draggable) {
-            if (!$(event.target).closest('section').is('#list-uncategorized-section')) {
-              GOVUK.curatedLists.postSort(event, draggable);
-            }
-
-            GOVUK.publishing.unlockPublishing();
-          },
-          receive: GOVUK.curatedLists.postSort
-        });
+      $('#all-list-items tr').draggable({
+        connectToSortable: '.drag-and-droppable',
+        helper: 'clone',
+        placeholder: 'droppable-placeholder'
       });
 
-      $('tr.untagged .remove').tooltip({
+      $('.drag-and-droppable').sortable({
+        connectWith: '.drag-and-droppable',
+        items: 'tr.item',
+        dropOnEmpty: true,
+        placeholder: 'droppable-placeholder',
+        update: function(event, _draggable) {
+          var $targetList = $(event.target);
+          GOVUK.curatedLists.reindexList($targetList);
+        }
+      });
+
+      $('tr.untagged .button_to').tooltip({
         title: "This item is no longer tagged with this tag, so it is no " +
         "longer visible to the user. You can safely remove it."
       })
-
     },
-    postSort: function(event, draggable) {
-      var $droppedRow = draggable.item;
-      var $targetList = $(event.target);
-      var $sourceList = draggable.sender;
-      var internalMove = ($droppedRow.data('list-id') === $targetList.data('list-id'));
+    reindexList: function($list) {
+      var $rows = $list.children(':not(.empty-list)');
 
-      $targetList.children('.empty-list').hide();
+      $rows.each(function(domIndex, row) {
+        var $row = $(row);
 
-      if ($targetList.closest('section').is('#list-uncategorized-section')) {
-        GOVUK.curatedLists.deleteRow($droppedRow);
-      } else {
-        var $allRows = $targetList.children(':not(.empty-list)');
+        // The item might have a new list ID.
+        $row.data({ 'list-id': $list.data('list-id') });
 
-        var startIndex = $droppedRow.data('index');
-        var stopIndex = $allRows.index($droppedRow);
+        var updateURL = $row.data('update-url');
 
-        var indexToUpdateFrom;
-        if ($sourceList) { // Reindexing destination list
-          $droppedRow.data('list-id', $targetList.data('list-id'));
-          indexToUpdateFrom = stopIndex;
-        } else if (internalMove) { // Move was within same list
-          indexToUpdateFrom = Math.min(startIndex, stopIndex);
-        } else { // Reindexing source list
-          indexToUpdateFrom = startIndex;
+        if (updateURL) {
+          GOVUK.curatedLists.updateRow(updateURL, $list, $row, domIndex);
+        } else {
+          GOVUK.curatedLists.createRow($list, $row, domIndex);
         }
+      });
 
-        var $rowsToUpdate = $allRows.slice(indexToUpdateFrom);
-
-        GOVUK.curatedLists.reindex($rowsToUpdate, indexToUpdateFrom);
-      }
-    },
-    reindex: function($rows, offset) {
-      if ($rows.length > 0) {
-        $rows.first().closest('section').css('opacity', '0.5');
-        var $list = $rows.first().closest('.curated-list');
-
-        $rows.each(function(domIndex, row) {
-          var $row = $(row);
-          var updateURL = $row.data('update-url');
-          var index = offset + domIndex;
-
-          $row.data('index', index);
-
-          if (updateURL) {
-            GOVUK.curatedLists.updateRow(updateURL, $list, $row, index);
-          } else {
-            GOVUK.curatedLists.createRow($list, $row, index);
-          }
-        });
-      } else {
-        $('.curated-list').each(function(_, list) {
-          if ($(list).children().not('.empty-list').length <= 0) {
-            $(list).children('.empty-list').show();
-          }
-        });
-      }
+      GOVUK.curatedLists.showEmptyLabels();
+      GOVUK.curatedLists.redrawUncategorizedList();
     },
     updateRow: function(updateURL, $list, $row, index) {
       $row.addClass('working');
@@ -103,7 +64,8 @@
         contentType: 'application/json',
         dataType: 'json',
         success: function() {
-          GOVUK.curatedLists.postReindex($list, $row);
+          $row.removeClass('working');
+          GOVUK.publishing.unlockPublishing();
         }
       });
     },
@@ -118,16 +80,16 @@
         data: JSON.stringify({
           list_item: {
             index: index,
-            title: $row.find('.title').text(),
-            api_url: $row.find('.api-url').text()
+            title: $row.data('title'),
+            api_url: $row.data('api-url')
           }
         }),
         contentType: 'application/json',
         dataType: 'json',
         success: function(data) {
           $row.removeClass('working');
-          $row.data('update-url', data['updateURL']);
-          GOVUK.curatedLists.postReindex($list);
+          $row.data({ 'update-url': data['updateURL'] });
+          GOVUK.publishing.unlockPublishing();
         }
       });
     },
@@ -136,23 +98,46 @@
 
       var deleteURL = $row.data('update-url');
 
+      if (!deleteURL) {
+        return;
+      }
+
       $.ajax(deleteURL, {
         type: 'DELETE',
         contentType: 'application/json',
         dataType: 'json',
         success: function() {
           $row.removeClass('working');
+          $row.remove();
+          GOVUK.curatedLists.redrawUncategorizedList();
+          GOVUK.publishing.unlockPublishing();
         }
       });
     },
-    postReindex: function($list, $row) {
-      if ($row) {
-        $row.removeClass('working');
-      }
-
-      if ($list.children('.working').length === 0) {
-        $list.closest('section').css('opacity', '1');
-      }
+    activateRemoveButtons: function () {
+      $('.curated-list').on('click', '.remove-button', function (e) {
+        e.preventDefault();
+        var $row = $(e.currentTarget).parent().parent().parent('tr');
+        GOVUK.curatedLists.deleteRow($row);
+        GOVUK.curatedLists.showEmptyLabels();
+        return false;
+      })
     },
+    showEmptyLabels: function () {
+      $('.curated-list').each(function(_, list) {
+        if ($(list).children().not('.empty-list').length <= 0) {
+          $(list).children('.empty-list').show();
+        } else {
+          $(list).children('.empty-list').hide();
+        }
+      });
+    },
+    redrawUncategorizedList: function () {
+      $('.is-curated').removeClass('is-curated');
+      $(".drag-and-droppable tr").each(function() {
+        var url = $(this).data('api-url')
+        $("[data-api-url='" + url + "']").addClass('is-curated');
+      })
+    }
   };
 }());
